Cleanup

+++ src/main/java/com/elmakers/mine/bukkit/action/BaseShopAction.java
@@ -168,7 +168,7 @@ public abstract class BaseShopAction extends BaseSpellAction implements GUIActio
         if (worth > 0) {
             if (isXP) {
                 worth = Math.ceil(costScale * worth * controller.getWorthBase() / controller.getWorthXP());
-                hasCosts = mage.getExperience() >= (int)(double)worth;
+                hasCosts = mage.getExperience() >= (int)worth;
             } else if (isItems) {
                 worth = Math.ceil(costScale * worth * controller.getWorthBase() / controller.getWorthItemAmount());
                 int hasAmount = getItemAmount(controller, mage);
@@ -193,7 +193,7 @@ public abstract class BaseShopAction extends BaseSpellAction implements GUIActio
 
         if (isXP) {
             worth = Math.ceil(costScale * worth * controller.getWorthBase() / controller.getWorthXP());
-            amountString = Integer.toString((int)(double)worth);
+            amountString = Integer.toString((int)worth);
             amountString = messages.get("costs.xp_amount").replace("$amount", amountString);
         }
         else if (isItems)
@@ -221,7 +221,7 @@ public abstract class BaseShopAction extends BaseSpellAction implements GUIActio
         double worth = shopItem.getWorth();
         if (isXP) {
             worth = Math.ceil(costScale * worth * controller.getWorthBase() / controller.getWorthXP());
-            mage.giveExperience((int) (double) worth);
+            mage.giveExperience((int) worth);
         }
         else if (isItems)
         {
@@ -255,7 +255,7 @@ public abstract class BaseShopAction extends BaseSpellAction implements GUIActio
 
         if (isXP) {
             worth = Math.ceil(costScale * worth * controller.getWorthBase() / controller.getWorthXP());
-            mage.removeExperience((int)(double)worth);
+            mage.removeExperience((int)worth);
         }
         else if (isItems)
         {
